- Modification du temps de simulation (double -> long) - Ajout de safe checks dans les setters (defensive programming) - Mise à jour des diagrammes uml - Correction d'un problème avec la configuration

+++ trunk/src/main/java/be/simulation/core/evenements/EventList.java
@@ -124,8 +124,7 @@ public class EventList {
 				// sinon inutile de le retourner.
 				if (i > 0) {
 					Evenement imminent = eventList.peekFirst();
-					if (tmp.getTempsPrevu().doubleValue() == imminent
-							.getTempsPrevu().doubleValue()) {
+					if (tmp.getTempsPrevu() == imminent.getTempsPrevu()) {
 						retVal = tmp;
 					}
 				} else {
